requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
pymysql==1.1.0
cryptography==41.0.5
openai==1.3.5  
httpx==0.24.1  
python-dotenv==1.0.0
reportlab==4.1.0
json5==0.9.22
gunicorn==21.2.0